server/requirements.txt
pynacl==1.5.0
flask==3.0.0